Improve documentation by adding a pseudocode section
This adds a paragraph to the projects READNE containing pseudocode
that illustrates the steps needed to decode an encodes polyline.

The main intent is to make it easier to understand how the
flexible poyline encoding algorithm works.
Potentially, it could also help in adding an implementaion in a
yet unsupported language.

Signed-off-by: Frank Gaebler <frank.gaebler@here.com>